third_party/blink/web_tests/platform/generic/inspector-protocol/css/css-set-inline-style-text-expected.txt

The test verifies functionality of protocol method CSS.setStyleTexts for inline elements.
==== Initial style sheet text ====
color: red;

Running test: testBasicSetStyle
==== Style sheet text ====
content: 'EDITED'
Stylesheet text: content: 'EDITED'

Running test: testSetStylePoorContent
Expected protocol error: Failed applying edit #0: SyntaxError Style text is not valid.
Stylesheet text: color: red;

Running test: testDeleteStyleBody
==== Style sheet text ====

Stylesheet text: 

Running test: testSetStyleOpenComment
Expected protocol error: Failed applying edit #0: SyntaxError Style text is not valid.
Stylesheet text: color: red;
